back_end/ai_audit_system/requirements.txt
pandas
pyyaml
fastembed
faiss-cpu
regex
numpy
tqdm
openpyxl

